fastapi_app/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
requests
python-multipart

